Support PTP `get_partialobject` to fix issue with large transfers
Large transfers, for example RAW files hitting 50 MiB, would cause a usb
error. I'm not too sure why but it seems like using smaller transfer
sizes fixes that. 15 MiB seemed like a sweet spot. I'm not sure if
that's dependent on the device I'm using which it may be. Works for now,
and for me tho.